+++ lua/startup.lua
@@ -17,7 +17,7 @@ CenterHelperCirclesEvery = centerHelperCirclesEvery or 1000
 CenterHelperCirclesEvery = uclamp(centerHelperCirclesEvery, 100, 2000)
 
 P("=========================================")
-P("YFS-Tools 1.7.1 (unofficial)")
+P("YFS-Tools 1.7.2 (unofficial)")
 P("created by tobitege (c) 2023")
 P("Thanks to Yoarii (SVEA) for YFS and LUA help!")
 P("YFS 1.4+ databank link required (Routes).")