+++ util/wolfeARlib.lua
@@ -1,5 +1,5 @@
--- wolflib.lua
--- Requires global WaypointInfo (Atlas data!)
+-- wolfeARlib.lua
+-- Requires stl.lua, global WaypointInfo (Atlas data!)
 
 local uclamp = utils.clamp
 
@@ -16,189 +16,6 @@ local font = 'Oxanium' -- 'Refrigerator'
 local colorBlue = { 130, 224, 255 } -- blue #66CCFF
 local colorLime = { 50, 205, 50 }   -- lime #32CD32
 local colorOrange = { 256, 128, 0 } -- orange #ff8000
-
-local SmartTemplateLibrary = (function ()
-    --[[
-      Wolfe Labs Smart Template Library (STL)
-      A simple, Twig-like templating language for Lua
-      Syntax:
-        {{ variable }} prints the contents of "variable"
-        {% some_lua_code %} executes the Lua code, useful for creating blocks like {% if %} and {% else %}, make sure you add {% end %} too :)
-      (C) 2022 - Wolfe Labs
-    ]]
-
-    --- Helper function that generates a clean print statement of a certain string
-    ---@param str string The string we need to show
-    ---@return string
-    local function mkPrint(str)
-      return 'print(\'' .. str:gsub('\'', '\\\''):gsub('\n', '\\n') .. '\')'
-    end
-
-    --- Helper function that merges tables
-    ---@vararg table
-    ---@return table
-    local function tMerge(...)
-      local tables = {...}
-      local result = {}
-      for _, t in pairs(tables) do
-        for k, v in pairs(t) do
-          result[k] = v
-        end
-      end
-      return result
-    end
-
-    ---@class Template
-    local Template = {
-      --- Globals available for every template by default
-      globals = {
-        math = math,
-        table = table,
-        string = string,
-        ipairs = ipairs,
-        pairs = pairs,
-      }
-    }
-
-    -- Makes our template directly callable
-    function Template.__call(self, ...)
-        local t = ...
-      return Template.render(self, ({...})[1])
-    end
-
-    --- Renders our template
-    ---@param vars table The variables to be used when rendering the template
-    ---@return string
-    function Template:render(vars)
-      -- Safety check, vars MUST be a table or nil
-      if type(vars or {}) ~= 'table' then
-        error('Template parameters must be a table, got ' .. type(vars))
-      end
-
-      --- This is our return buffer
-      local _ = {}
-
-      -- Creates our environment
-      local env = tMerge(Template.globals, self.globals or {}, vars or {}, {
-        print = function (str) table.insert(_, tostring(str or '')) end,
-      })
-
-      -- Invokes our template
-      self.callable(env)
-
-      -- General trimming
-      local result = table.concat(_, ''):gsub('%s+', ' ')
-
-      -- Trims result
-      result = result:sub(result:find('[^%s]') or 1):gsub('%s*$', '')
-
-      -- Done
-      return result
-    end
-
-    --- Creates a new template
-    ---@param source string The code for your template
-    ---@param globals table Global variables to be used on on the template
-    ---@param buildErrorHandler function A function to handle build errors, if none is found throws an error
-    ---@return Template|nil
-    function Template.new(source, globals, buildErrorHandler)
-      -- Creates our instance
-      local self = {
-        source = source,
-        globals = globals,
-      }
-
-      -- Yield function (mostly for games who limit executions per frame)
-      local yield = (coroutine and coroutine.isyieldable() and coroutine.yield) or function () end
-
-      -- Parses direct printing of variables, we'll convert a {{var}} into {% print(var) %}
-      source = source:gsub('{{(.-)}}', '{%% print(%1) %%}')
-
-      -- Ensures {% if %} ... {% else %} ... {% end %} stays on same line
-      source = source:gsub('\n%s*{%%', '{%%')
-      source = source:gsub('%%}\n', '%%}')
-
-      --- This variable stores all our Lua "pieces"
-      local tPieces = {}
-
-      -- Parses actual Lua inside {% lua %} tags
-      while #source > 0 do
-        --- The start index of Lua tag
-        local iLuaStart = source:find('{%%')
-
-        --- The end index of Lua tag
-        local iLuaEnd = source:find('%%}')
-
-        -- Checks if we have a match
-        if iLuaStart then
-          -- Errors when not closing a tag
-          if not iLuaEnd then
-            error('Template error, missing Lua closing tag near: ' .. source:sub(0, 16))
-          end
-
-          --- The current text before Lua tag
-          local currentText = source:sub(1, iLuaStart - 1)
-          if #currentText then
-            table.insert(tPieces, mkPrint(currentText))
-          end
-
-          --- Our Lua tag content
-          local luaTagContent = source:sub(iLuaStart, iLuaEnd + 1):match('{%%(.-)%%}') or ''
-          table.insert(tPieces, luaTagContent)
-
-          -- Removes parsed content
-          source = source:sub(iLuaEnd + 2)
-        else
-          -- Adds remaining Lua as a single print statement
-          table.insert(tPieces, mkPrint(source))
-
-          -- Marks content as parsed
-          source = ''
-        end
-
-        -- Yields loading
-        yield()
-      end
-
-      -- Builds the Lua function
-      self.code = table.concat(tPieces, '\n')
-
-      -- Builds our function and caches it, this is our template now
-      local _, err = load(string.format([[return function (_) _ENV = _; _ = _ENV[_]; %s; end]], self.code), nil, 't', {})
-      if _ and not err then
-        _ = _()
-      end
-
-      -- Checks for any errors
-      if err then
-        if buildErrorHandler then
-          buildErrorHandler(self, err)
-        else
-          error('[E] Failed compiling template: ' .. err)
-        end
-
-        -- Retuns an invalid instance
-        return nil
-      else
-        -- If everything passed, assigns our callable to our compiled function
-        self.callable = _
-      end
-
-      -- Initializes our instance
-      return setmetatable(self, Template)
-    end
-
-    -- By default, returns the constructor of our class
-    return Template.new
-end)()
-
-local function SmartTemplate(code, globals)
-return SmartTemplateLibrary(code, globals, function(template, err)
-    DUSystem.print(' [ERROR] Failed compiling template: ' .. err)
-    DUSystem.print('[SOURCE] ' .. code)
-    error()
-end)
-end
 
 --- Distance to "infinity", used when projecting AR directions, set to 100su
 ---@type number
@@ -369,14 +186,14 @@ function wolfie.render(destList)
     local Shapes = {}
 
     --- Draws the hexagon, showing the current destiantion point in space
-    Shapes.Hexagon = SmartTemplate([[
+    Shapes.Hexagon = STL([[
     <svg style="width: {{ size or 1 }}em; height: {{ size or 1 }}em;" viewBox="0 0 48 48" fill="none" xmlns="http://www.w3.org/2000/svg">
         <path d="M24.25 1.56699L24 1.42265L23.75 1.56699L4.69745 12.567L4.44745 12.7113V13V35V35.2887L4.69745 35.433L23.75 46.433L24 46.5774L24.25 46.433L43.3026 35.433L43.5526 35.2887V35V13V12.7113L43.3026 12.567L24.25 1.56699ZM9.44745 32.4019V15.5981L24 7.19615L38.5526 15.5981V32.4019L24 40.8038L9.44745 32.4019Z" fill="{{ color }}" stroke="{{ stroke }}"/>
     </svg>
     ]])
 
     --- Draws the crosshair, showing current forward direction
-    Shapes.Crosshair = SmartTemplate([[
+    Shapes.Crosshair = STL([[
     <svg style="width: {{ size or 1 }}em; height: {{ size or 1 }}em;" viewBox="0 0 48 48" fill="none" xmlns="http://www.w3.org/2000/svg">
         <path d="M23.6465 37.8683L24.0001 38.2218L24.3536 37.8683L26.3536 35.8684L26.5 35.7219V35.5148V26.5H35.5148H35.7219L35.8684 26.3536L37.8684 24.3536L38.2219 24L37.8684 23.6465L35.8684 21.6465L35.7219 21.5H35.5148H26.5V12.4852V12.2781L26.3536 12.1317L24.3536 10.1317L24.0001 9.77818L23.6465 10.1317L21.6465 12.1318L21.5 12.2782V12.4854V21.5H12.4854H12.2782L12.1318 21.6465L10.1318 23.6465L9.77824 24L10.1318 24.3536L12.1318 26.3536L12.2782 26.5H12.4854H21.5V35.5147V35.7218L21.6465 35.8682L23.6465 37.8683Z" fill="{{ color }}" stroke="{{ stroke }}"/>
     </svg>
@@ -404,7 +221,7 @@ function wolfie.render(destList)
 
     --- Draws text
     -- text-shadow: 0px 0px 0.125em {{ stroke or Colors.Shadow }}, 0px 0px 0.250em #000, 0px 0px 0.500em #000;">
-    renderGlobals.Label = SmartTemplate([[<span style="font-size: {{ size or ']]..fontSize..[[' }}em; font-family: {{ font or ']]..font..[[' }};
+    renderGlobals.Label = STL([[<span style="font-size: {{ size or ']]..fontSize..[[' }}em; font-family: {{ font or ']]..font..[[' }};
         font-weight: {{ weight or 'normal' }}; color: {{ color or GetHudColor() }};
         text-shadow: 2px 2px 0.125em {{ stroke or Colors.Shadow }}, -2px -2px 0.50em #000, 2px -2px 0.50em #000;">
     {{ text }}
@@ -412,11 +229,11 @@ function wolfie.render(destList)
     ]], renderGlobals)
 
     --- Creates an element that is always centered at a certain coordinate
-    UI.PositionCenteredAt = SmartTemplate('position: absolute; top: {{ Percentage(y, 5) }}; left: {{ Percentage(x, 5) }}; margin-top: -{{ (height or 1) / 2 }}em; margin-left: -{{ (width or 1) / 2 }}em;'
+    UI.PositionCenteredAt = STL('position: absolute; top: {{ Percentage(y, 5) }}; left: {{ Percentage(x, 5) }}; margin-top: -{{ (height or 1) / 2 }}em; margin-left: -{{ (width or 1) / 2 }}em;'
     , renderGlobals)
 
     --- Renders a full destination marker (hexagon + info)
-    UI.DestinationMarker = SmartTemplate([[
+    UI.DestinationMarker = STL([[
     {%
     local screen = WorldCoordinate(position)
     local distance = DistanceTo(position)
@@ -455,7 +272,7 @@ function wolfie.render(destList)
     ]], renderGlobals)
 
     --- Renders a crosshair shape
-    UI.Crosshair = SmartTemplate([[
+    UI.Crosshair = STL([[
     <div style="{{ UI.PositionCenteredAt({ x = x, y = y, width = 1.5, height = 1.5 }) }}">
         {{ Shapes.Crosshair({ color = GetHudColor(), stroke = Colors.Shadow, size = 1.5 }) }}
     </div>
@@ -492,7 +309,7 @@ function wolfie.render(destList)
                 isDestinationOnSameCelestialBody = true
             end
 
-            local marker = SmartTemplate([[ {{ UI.DestinationMarker({ title = currentDestination.name, position = currentDestination.position }) }} ]],
+            local marker = STL([[ {{ UI.DestinationMarker({ title = currentDestination.name, position = currentDestination.position }) }} ]],
             {
                 currentDestination = dest,
                 currentCelestialBody = currentCelestialBodyInfo,
@@ -505,7 +322,7 @@ function wolfie.render(destList)
         end
     end
     rAR = rAR .. "</div>"
-    local renderAR = SmartTemplate(rAR, renderGlobals)
+    local renderAR = STL(rAR, renderGlobals)
     if renderAR then
         return renderAR()
     end